Data.Bytes

Description

This module serves as the main entry point for the library. It provides the types and operations for typical usage and is usually the only import required. The core concept is:

Wrapping a numeric value representing bytes in a new type.

Attaching phantom labels representing the units (e.g. K, M, ...).

This prevents mistakes, such as adding two different byte sizes or converting between sizes incorrectly.

Since: 0.1

Types

Units

data Size

Byte units.

Since: 0.1

Constructors

B
Bytes

Since: 0.1

K
Kilobytes

Since: 0.1

M
Megabytes

Since: 0.1

G
Gigabytes

Since: 0.1

T
Terabytes

Since: 0.1

P
Petabytes

Since: 0.1

E
Exabytes

Since: 0.1

Z
Zettabytes

Since: 0.1

Y
Yottabytes

Since: 0.1

class Sized a where

Types that have a size.

Examples

Expand

>>> import Data.Bytes
>>> sizeOf (MkBytes @G 7)
G

>>> sizeOf (hideSize $ MkBytes @M 7)
M

Since: 0.1

Methods

sizeOf :: a -> Size

Retrieves the size.

Since: 0.1

Bytes

newtype Bytes (s :: Size) (n :: Type)

This is the core type for handling type-safe byte operations. It is intended to be used as a simple wrapper over some numeric type, equipped with a Size tag.

To take full advantage of the API (e.g. normalize), the underlying numeric type should implement Semifield or, ideally, Field.

Examples

Expand

>>> MkBytes @M 1000
MkBytes 1000

Since: 0.1

Constructors

MkBytes n

class Unwrapper a where

Abstracts "unwrapper" types for generically retrieving a wrapped value.

>>> import Data.Bytes
>>> unwrap (MkBytes @G 7)
7

>>> unwrap (hideSize $ MkBytes @M 400)
400

Since: 0.1

Associated Types

type Unwrapped a

Since: 0.1

Methods

unwrap :: a -> Unwrapped a

Retrieves the underlying type.

Since: 0.1

Unknown Size

data SomeSize (n :: Type)

Wrapper for Bytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getFileSize :: FilePath -> IO (SomeSize Float)
  getFileSize path = do
    -- getRawFileSize :: FilePath -> IO (Float, String)
    (bytes, units) <- getRawFileSize path
    pure $ case units of
      "B" -> hideSize $ MkBytes @B bytes
      "K" -> hideSize $ MkBytes @K bytes
      _ -> error "todo"
:}

SomeSize carries along an SSize runtime witness for when we need the size. Its Algebra functions are normalized.

We define an equivalence relation on SomeSize that takes units into account. For instance,

>>> hideSize (MkBytes @G 7) == hideSize (MkBytes @M 7_000)
True

Because we expose the underlying Bytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

hideSize :: forall s n. SingSize s => Bytes s n -> SomeSize n

Wraps a Bytes in an existentially quantified SomeSize.

Since: 0.1

Transformations

Converting Units

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = toG bytes
>>> :type gBytes
gBytes :: Bytes 'G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = toM bytes
>>> :type mBytes
mBytes :: Bytes 'M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Associated Types

type Converted (b :: Size) a = r | r -> b

Since: 0.1

Methods

toB :: a -> Converted B a

Since: 0.1

toK :: a -> Converted K a

Since: 0.1

toM :: a -> Converted M a

Since: 0.1

toG :: a -> Converted G a

Since: 0.1

toT :: a -> Converted T a

Since: 0.1

toP :: a -> Converted P a

Since: 0.1

toE :: a -> Converted E a

Since: 0.1

toZ :: a -> Converted Z a

Since: 0.1

toY :: a -> Converted Y a

Since: 0.1

Normalization

class Normalize a where

Used for normalizing bytes b such that

\[ 1 \le \text{normalize}(b) < 1000 \iff 1\text{ B} \le b < 1000\text{ Y}. \]

In the strictest sense, \(\textrm{normalize} : \mathcal{C} \rightarrow \mathcal{C}\) is not a homomorphism, as the combination of two normalized values may itself not be normalized.

However, because the normalized units varies with the value, normalize always returns a type that existentially quantifies the size (e.g. SomeSize). Eq for these types is defined in terms of an equivalence class that takes units into account e.g. 1 P = 1,000 T = 1,000,000 G .... Viewed this way, normalize is actually an isomorphism, as it is essentially a no-op, never leaving the equivalence class.

This means we can happily mix normalization with different functions without worrying about the order. The only requirement we have is that such functions respect substitution:

\[ x = y \implies f(x) = f(y). \]

This is certainly true for all the usual mathematical operations we would normally use, e.g., AGroup addition, Module scalar multiplication. On the other hand, any functions that inspect the underlying numeric value or Bytes types could easily break this law. As such they should be treated with suspicion, at least when used in conjunction with normalize.

The other consideration we must keep in mind is that the final result of a series of computations may not be normalized. If this is desired, then normalize should be the last operation performed. Using normalize in the middle would not cause any harm (other than, perhaps, impacting efficiency), but it would not guarantee the final result is normalized.

Examples

Expand

>>> let bytes = MkBytes 5000 :: Bytes 'M Int
>>> normalize bytes
MkSomeSize SG (MkBytes 5)

>>> let bytes = hideSize (MkBytes 0.01 :: Bytes 'T Float)
>>> normalize bytes
MkSomeSize SG (MkBytes 10.0)

Since: 0.1

Associated Types

type Norm a

Since: 0.1

Methods

normalize :: a -> Norm a

Since: 0.1

Algebra

The built-in Num class is abandoned in favor of algebra-simple's algebraic hierarchy based on abstract algebra. This is motivated by a desire to:

Provide a consistent API.

Avoid Num's infelicities (e.g. nonsense multiplication, dangerous fromInteger).

Bytes and SomeSize are both AGroups. A Ring instance is not provided because multiplication is nonsensical:

\[ x \;\textrm{mb} \times y \;\textrm{mb} = xy \;\textrm{mb}^2. \]

Fortunately, multiplying bytes by some kind of scalar is both useful and has an easy interpretation: Bytes forms a Module over a Ring (resp. VectorSpace over a Field). This allows us to multiply a Bytes or SomeSize by a scalar in a manner consistent with the above API.

Examples

Addition/Subtraction

>>> import Numeric.Algebra (ASemigroup ((.+.)), AGroup ((.-.)))
>>> let mb1 = MkBytes 20 :: Bytes 'M Int
>>> let mb2 = MkBytes 50 :: Bytes 'M Int
>>> mb1 .+. mb2
MkBytes 70
>>> mb1 .-. mb2
MkBytes (-30)

>>> let kb = MkBytes 50 :: Bytes 'K Int
>>> -- mb1 .+. kb -- This would be a type error

Multiplication

>>> import Numeric.Algebra (MSemiSpace ((.*)))
>>> mb1 .* 10
MkBytes 200

Division

>>> import Numeric.Algebra (MSpace ((.%)))
>>> import Numeric.Data.NonZero (unsafeNonZero)
>>> mb1 .% (unsafeNonZero 10)
MkBytes 2

One may wonder how the AGroup instance for SomeSize could possibly work. It is possible (indeed, expected) that we could have two SomeSizes that have different underlying Bytes types. To handle this, the SomeSize instance will convert both Bytes to a Bytes 'B before adding/subtracting.

>>> let some1 = hideSize (MkBytes 1000 :: Bytes 'G Double)
>>> let some2 = hideSize (MkBytes 500_000 :: Bytes 'M Double)
>>> some1 .+. some2
MkSomeSize SB (MkBytes 1.5e12)
>>> some1 .-. some2
MkSomeSize SB (MkBytes 5.0e11)

This respects SomeSize's equivalence-class based Eq.

Text

Pretty Printing

We provide several formatters for pretty-printing different byte types.

>>> import Data.Default (Default (def))
>>> let bf = MkFloatingFormatter (Just 2)
>>> let b = MkBytes @G @Float 20.248
>>> formatSized bf def b
"20.25 gb"

module Data.Bytes.Formatting

Parsing

We provide tools for parsing byte types from Text. Parsing is lenient in general. We support:

Case-insensitivity.

Optional leading/internal/trailing whitespace.

Flexible names.

Examples

>>> parse @(Bytes M Int) "70"
Right (MkBytes 70)

>>> parse @(SomeSize Float) "100.45 kilobytes"
Right (MkSomeSize SK (MkBytes 100.45))

>>> parse @(SomeSize Word) "2300G"
Right (MkSomeSize SG (MkBytes 2300))

>>> parse @(SomeSize Float) "5.5 tb"
Right (MkSomeSize ST (MkBytes 5.5))

parse :: Parser a => Text -> Either Text a

Runs a Parser.

Since: 0.1

Reexports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.