Data.Bytes.Class.Conversion

Description

Provides typeclasses for convert between byte sizes.

Since: 0.1

Documentation

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = toG bytes
>>> :type gBytes
gBytes :: Bytes 'G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = toM bytes
>>> :type mBytes
mBytes :: Bytes 'M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Associated Types

type Converted (b :: Size) a = r | r -> b

Since: 0.1

Methods

toB :: a -> Converted B a

Since: 0.1

toK :: a -> Converted K a

Since: 0.1

toM :: a -> Converted M a

Since: 0.1

toG :: a -> Converted G a

Since: 0.1

toT :: a -> Converted T a

Since: 0.1

toP :: a -> Converted P a

Since: 0.1

toE :: a -> Converted E a

Since: 0.1

toZ :: a -> Converted Z a

Since: 0.1

toY :: a -> Converted Y a

Since: 0.1

convert :: forall n. (MGroup n, NumLiteral n) => Size -> Size -> n -> n

Low level function for converting a numeric literal between byte sizes. convert b1 b2 converts from b1 to b2, e.g. convert G K = \n -> n * 1_000_000. The higher level byte types and functions should be preferred (e.g. Conversion, Data.Bytes.Class.Normalize), but this is here when it is needed.

Since: 0.1

convertWitness :: forall s n. (MGroup n, NumLiteral n, SingSize s) => Size -> n -> n

Low level function for converting a numeric literal from the inferred SingSize to the parameter Size. For instance,

>>> convertWitness @K M 5_000
5

This is slightly more principled than convert, but the higher level byte types and functions should still be preferred (e.g. Bytes, Normalize).

Since: 0.1