Data.Bytes.Internal

Description

Internal module for Data.Bytes. The primary difference is this module exposes some underlying details i.e. singleton witnesses. These are hidden by default as they complicate the API.

Since: 0.1

Bytes

newtype Bytes (s :: Size) (n :: Type)

This is the core type for handling type-safe byte operations. It is intended to be used as a simple wrapper over some numeric type, equipped with a Size tag.

To take full advantage of the API (e.g. normalize), the underlying numeric type should implement Semifield or, ideally, Field.

Examples

Expand

>>> MkBytes @M 1000
MkBytes 1000

Since: 0.1

Constructors

MkBytes n

bytesToSSize :: SingSize s => Bytes s n -> SSize s

Retrieves the SSize witness. Can be used to recover the Size.

>>> bytesToSSize (MkBytes @K @Int 7)
SK

Since: 0.1

Unknown Size

data SomeSize (n :: Type) where

Wrapper for Bytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getFileSize :: FilePath -> IO (SomeSize Float)
  getFileSize path = do
    -- getRawFileSize :: FilePath -> IO (Float, String)
    (bytes, units) <- getRawFileSize path
    pure $ case units of
      "B" -> hideSize $ MkBytes @B bytes
      "K" -> hideSize $ MkBytes @K bytes
      _ -> error "todo"
:}

SomeSize carries along an SSize runtime witness for when we need the size. Its Algebra functions are normalized.

We define an equivalence relation on SomeSize that takes units into account. For instance,

>>> hideSize (MkBytes @G 7) == hideSize (MkBytes @M 7_000)
True

Because we expose the underlying Bytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

Constructors

MkSomeSize :: SSize s -> Bytes s n -> SomeSize n
Since: 0.1

hideSize :: forall s n. SingSize s => Bytes s n -> SomeSize n

Wraps a Bytes in an existentially quantified SomeSize.

Since: 0.1