Data.Bytes.Network

Description

This module serves as an alternative entry point to Data.Bytes, for when tracking uploaded vs. downloaded bytes is necessary. It provides the types and operations for typical usage and is usually the only import required. The core concept is:

Wrapping a numeric value representing bytes in a new type.

Attaching phantom labels representing the units (e.g. B, K, M, ...).

Attaching phantom labels representing the direction (i.e. Down, Up).

This prevents mistakes, such as adding two different byte sizes/directions, or converting between sizes incorrectly.

Since: 0.1

Types

Units

data Size

Byte units.

Since: 0.1

Constructors

B
Bytes

Since: 0.1

K
Kilobytes

Since: 0.1

M
Megabytes

Since: 0.1

G
Gigabytes

Since: 0.1

T
Terabytes

Since: 0.1

P
Petabytes

Since: 0.1

E
Exabytes

Since: 0.1

Z
Zettabytes

Since: 0.1

Y
Yottabytes

Since: 0.1

class Sized a where

Types that have a size.

Examples

Expand

>>> import Data.Bytes
>>> sizeOf (MkBytes @G 7)
G

>>> sizeOf (hideSize $ MkBytes @M 7)
M

Since: 0.1

Methods

sizeOf :: a -> Size

Retrieves the size.

Since: 0.1

data Direction

Tags for differentiating downloaded vs. uploaded bytes.

Since: 0.1

Constructors

Down
Since: 0.1

Up
Since: 0.1

class Directed a where

Types that have a direction.

Examples

Expand

>>> import Data.Bytes.Network
>>> directionOf (MkNetBytesP @Up @G 7)
Up

>>> directionOf (hideNetSizeDir $ MkNetBytesP @Down @M 100)
Down

Since: 0.1

Methods

directionOf :: a -> Direction

Retrieve the direction.

Since: 0.1

Bytes

newtype NetBytes (d :: Direction) (s :: Size) (n :: Type)

Wrapper around the Bytes type that adds the Direction tag.

Examples

Expand

>>> MkNetBytesP @Up @M 1000
MkNetBytes (MkBytes 1000)

Since: 0.1

Constructors

MkNetBytes (Bytes s n)

Bundled Patterns

pattern MkNetBytesP :: forall d s n. n -> NetBytes d s n
Pattern for de/constructing NetBytes.

Since: 0.1

class Unwrapper a where

Abstracts "unwrapper" types for generically retrieving a wrapped value.

>>> import Data.Bytes
>>> unwrap (MkBytes @G 7)
7

>>> unwrap (hideSize $ MkBytes @M 400)
400

Since: 0.1

Associated Types

type Unwrapped a

Since: 0.1

Methods

unwrap :: a -> Unwrapped a

Retrieves the underlying type.

Since: 0.1

Unknown Size

data SomeNetSize (d :: Direction) (n :: Type)

Wrapper for NetBytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getUpTraffic :: IO (SomeNetSize Up Float)
  getUpTraffic = do
    -- getUpTrafficRaw :: IO (Float, String)
    (bytes, units) <- getUpTrafficRaw
    pure $ case units of
      "B" -> hideNetSize $ MkNetBytesP @Up @B bytes
      "K" -> hideNetSize $ MkNetBytesP @Up @K bytes
      _ -> error "todo"
:}

SomeNetSize carries along an SSize runtime witness for when we need the size. Its Algebra functions are normalized.

We define an equivalence relation on SomeNetSize that takes units into account. For instance,

>>> hideNetSize (MkNetBytesP @Up @K 1000) == hideNetSize (MkNetBytesP @Up @M 1)
True

Because we expose the underlying NetBytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

hideNetSize :: forall d s n. SingSize s => NetBytes d s n -> SomeNetSize d n

Wraps a NetBytes in an existentially quantified SomeNetSize.

Since: 0.1

Unknown Direction

data SomeNetDir (s :: Size) (n :: Type)

Wrapper for NetBytes, existentially quantifying the direction. This is useful when a function does not know a priori what direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNetDir K Double)
  getMaxTraffic = do
    -- getMaxTrafficKRaw :: IO (Double, String)
    (bytes, direction) <- getMaxTrafficKRaw
    pure $ case direction of
      "down" -> hideNetDir $ MkNetBytesP @Down bytes
      "up" -> hideNetDir $ MkNetBytesP @Up bytes
      _ -> error "bad direction"
:}

We deliberately do not provide instances for SomeX classes that could be used to combine arbitrary SomeNetDirs (e.g. Applicative, AGroup), as that would defeat the purpose of enforcing the distinction between upload and downloaded bytes.

Equality is determined by the usual equivalence class -- that takes units into account -- and by considering the direction.

>>> hideNetDir (MkNetBytesP @Up @K 1000) == hideNetDir (MkNetBytesP @Up @K 1000)
True
>>> hideNetDir (MkNetBytesP @Up @K 1000) /= hideNetDir (MkNetBytesP @Down @K 1000)
True

Notice no Ord instance is provided, as we provide no ordering for Direction.

Since: 0.1

hideNetDir :: forall d s n. SingDirection d => NetBytes d s n -> SomeNetDir s n

Wraps a NetBytes in an existentially quantified SomeNetDir.

Since: 0.1

Unknown Size and Direction

data SomeNet (n :: Type)

Wrapper for NetBytes, existentially quantifying the size and direction. This is useful when a function does not know a priori what size or direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNet Double)
  getMaxTraffic = do
    -- getMaxTrafficRaw :: IO (Double, String, String)
    (bytes, direction, size) <- getMaxTrafficRaw
    pure $ case (direction, size) of
      ("Down", "K") -> hideNetSizeDir $ MkNetBytesP @Down @K bytes
      ("Up", "M") -> hideNetSizeDir $ MkNetBytesP @Up @M bytes
      _ -> error "todo"
:}

SomeNet carries along SDirection and SSize runtime witnesses for recovering the Direction and Size, respectively.

This instance uses the same equivalence relation from SomeNetSize w.r.t the size, and also includes an equality check on the direction. Thus we have, for instance,

>>> hideNetSizeDir (MkNetBytesP @Up @K 1_000) == hideNetSizeDir (MkNetBytesP @Up @M 1)
True

>>> hideNetSizeDir (MkNetBytesP @Up @K 1_000) /= hideNetSizeDir (MkNetBytesP @Down @M 1)
True

Since: 0.1

hideNetSizeDir :: forall d s n. (SingDirection d, SingSize s) => NetBytes d s n -> SomeNet n

Wraps a NetBytes in an existentially quantified SomeNet.

Since: 0.1

Transformations

Converting Units

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = toG bytes
>>> :type gBytes
gBytes :: Bytes 'G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = toM bytes
>>> :type mBytes
mBytes :: Bytes 'M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Associated Types

type Converted (b :: Size) a = r | r -> b

Since: 0.1

Methods

toB :: a -> Converted B a

Since: 0.1

toK :: a -> Converted K a

Since: 0.1

toM :: a -> Converted M a

Since: 0.1

toG :: a -> Converted G a

Since: 0.1

toT :: a -> Converted T a

Since: 0.1

toP :: a -> Converted P a

Since: 0.1

toE :: a -> Converted E a

Since: 0.1

toZ :: a -> Converted Z a

Since: 0.1

toY :: a -> Converted Y a

Since: 0.1

Normalization

class Normalize a where

Used for normalizing bytes b such that

\[ 1 \le \text{normalize}(b) < 1000 \iff 1\text{ B} \le b < 1000\text{ Y}. \]

In the strictest sense, \(\textrm{normalize} : \mathcal{C} \rightarrow \mathcal{C}\) is not a homomorphism, as the combination of two normalized values may itself not be normalized.

However, because the normalized units varies with the value, normalize always returns a type that existentially quantifies the size (e.g. SomeSize). Eq for these types is defined in terms of an equivalence class that takes units into account e.g. 1 P = 1,000 T = 1,000,000 G .... Viewed this way, normalize is actually an isomorphism, as it is essentially a no-op, never leaving the equivalence class.

This means we can happily mix normalization with different functions without worrying about the order. The only requirement we have is that such functions respect substitution:

\[ x = y \implies f(x) = f(y). \]

This is certainly true for all the usual mathematical operations we would normally use, e.g., AGroup addition, Module scalar multiplication. On the other hand, any functions that inspect the underlying numeric value or Bytes types could easily break this law. As such they should be treated with suspicion, at least when used in conjunction with normalize.

The other consideration we must keep in mind is that the final result of a series of computations may not be normalized. If this is desired, then normalize should be the last operation performed. Using normalize in the middle would not cause any harm (other than, perhaps, impacting efficiency), but it would not guarantee the final result is normalized.

Examples

Expand

>>> let bytes = MkBytes 5000 :: Bytes 'M Int
>>> normalize bytes
MkSomeSize SG (MkBytes 5)

>>> let bytes = hideSize (MkBytes 0.01 :: Bytes 'T Float)
>>> normalize bytes
MkSomeSize SG (MkBytes 10.0)

Since: 0.1

Associated Types

type Norm a

Since: 0.1

Methods

normalize :: a -> Norm a

Since: 0.1

Algebra

The built-in Num class is abandoned in favor of algebra-simple's algebraic hierarchy based on abstract algebra. This is motivated by a desire to:

Provide a consistent API.

Avoid Num's infelicities (e.g. nonsense multiplication, dangerous fromInteger).

NetBytes and SomeNetSize are both AGroups. A Ring instance is not provided because multiplication is nonsensical:

\[ x \;\textrm{mb} \times y \;\textrm{mb} = xy \;\textrm{mb}^2. \]

Fortunately, multiplying bytes by some kind of scalar is both useful and has an easy interpretation: NetBytes forms a Module over a Ring (resp. VectorSpace over a Field). This allows us to multiply a NetBytes or SomeNetSize by a scalar in a manner consistent with the above API.

Examples

Addition/Subtraction

>>> import Numeric.Algebra (ASemigroup ((.+.)), AGroup ((.-.)))
>>> let mb1 = MkNetBytesP 20 :: NetBytes 'Down 'M Int
>>> let mb2 = MkNetBytesP 50 :: NetBytes 'Down 'M Int
>>> mb1 .+. mb2
MkNetBytes (MkBytes 70)
>>> mb1 .-. mb2
MkNetBytes (MkBytes (-30))

>>> let kb = MkNetBytesP 50 :: NetBytes 'Down 'K Int
>>> -- mb1 .+. kb -- This would be a type error

>>> let mbUp = MkNetBytesP 50 :: NetBytes 'Up 'M Int
>>> -- mb1 .+. mbUp -- This would be a type error

Multiplication

>>> import Numeric.Algebra (MSemiSpace ((.*)))
>>> mb1 .* 10
MkNetBytes (MkBytes 200)

Division

>>> import Numeric.Algebra (MSpace ((.%)))
>>> import Numeric.Data.NonZero (unsafeNonZero)
>>> mb1 .% (unsafeNonZero 10)
MkNetBytes (MkBytes 2)

One may wonder how the AGroup instance for SomeNetSize could possibly work. It is possible (indeed, expected) that we could have two SomeNetSizes that have different underlying NetBytes types. To handle this, the SomeNetSize instance will convert both NetBytes to a NetBytes 'B before adding/subtracting.

>>> let some1 = hideNetSize (MkNetBytesP 1000 :: NetBytes 'Down 'G Double)
>>> let some2 = hideNetSize (MkNetBytesP 500_000 :: NetBytes 'Down 'M Double)
>>> some1 .+. some2
MkSomeNetSize SB (MkNetBytes (MkBytes 1.5e12))
>>> some1 .-. some2
MkSomeNetSize SB (MkNetBytes (MkBytes 5.0e11))

This respects SomeNetSize's equivalence-class base Eq.

Text

Pretty Printing

We provide several formatters for pretty-printing different byte types.

>>> import Data.Default (Default (def))
>>> let bf = MkFloatingFormatter (Just 2)
>>> let b = MkNetBytesP @Up @G @Float 203.301
>>> formatSizedDirected bf def def b
"203.30 gb up"

module Data.Bytes.Formatting

Parsing

We provide tools for parsing byte types from Text. Parsing is lenient in general. We support:

Case-insensitivity.

Optional leading/internal/trailing whitespace. Note the there must be at least some whitespace between size and direction units.

Flexible names.

Examples

>>> parse @(NetBytes Up M Int) "70"
Right (MkNetBytes (MkBytes 70))

>>> parse @(SomeNetSize Down Float) "100.45 kilobytes"
Right (MkSomeNetSize SK (MkNetBytes (MkBytes 100.45)))

>>> parse @(SomeNetSize Up Word) "2300G"
Right (MkSomeNetSize SG (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir T Word) "2300 up"
Right (MkSomeNetDir SUp (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir M Word) "2300D"
Right (MkSomeNetDir SDown (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNet Float) "5.5 tb Up"
Right (MkSomeNet SUp ST (MkNetBytes (MkBytes 5.5)))

>>> parse @(SomeNet Float) "5.5 megabytes DOWN"
Right (MkSomeNet SDown SM (MkNetBytes (MkBytes 5.5)))

parse :: Parser a => Text -> Either Text a

Runs a Parser.

Since: 0.1

Reexports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.