Data.Bytes.Network.NetBytes

Description

Provides the NetBytes and SomeNetSize types as an alternative to Data.Bytes, for when there is a need to distinguish between downloaded and uploaded bytes.

Since: 0.1

Network Bytes

newtype NetBytes (d :: Direction) (s :: Size) (n :: Type)

Wrapper around the Bytes type that adds the Direction tag.

Examples

Expand

>>> MkNetBytesP @Up @M 1000
MkNetBytes (MkBytes 1000)

Since: 0.1

Constructors

MkNetBytes (Bytes s n)

Bundled Patterns

pattern MkNetBytesP :: forall d s n. n -> NetBytes d s n
Pattern for de/constructing NetBytes.

Since: 0.1

hideNetSize :: forall d s n. SingSize s => NetBytes d s n -> SomeNetSize d n

Wraps a NetBytes in an existentially quantified SomeNetSize.

Since: 0.1

Unknown Size

data SomeNetSize (d :: Direction) (n :: Type)

Wrapper for NetBytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getUpTraffic :: IO (SomeNetSize Up Float)
  getUpTraffic = do
    -- getUpTrafficRaw :: IO (Float, String)
    (bytes, units) <- getUpTrafficRaw
    pure $ case units of
      "B" -> hideNetSize $ MkNetBytesP @Up @B bytes
      "K" -> hideNetSize $ MkNetBytesP @Up @K bytes
      _ -> error "todo"
:}

SomeNetSize carries along an SSize runtime witness for when we need the size. Its Algebra functions are normalized.

We define an equivalence relation on SomeNetSize that takes units into account. For instance,

>>> hideNetSize (MkNetBytesP @Up @K 1000) == hideNetSize (MkNetBytesP @Up @M 1)
True

Because we expose the underlying NetBytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1