Data.Bytes.Network.SomeNetDir.Internal

Description

Internal module for Data.Bytes.Network.SomeNetDir. The primary difference is this module exposes some underlying details i.e. singleton witnesses. These are hidden by default as they complicate the API.

Since: 0.1

Unknown Direction

data SomeNetDir (s :: Size) (n :: Type) where

Wrapper for NetBytes, existentially quantifying the direction. This is useful when a function does not know a priori what direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNetDir K Double)
  getMaxTraffic = do
    -- getMaxTrafficKRaw :: IO (Double, String)
    (bytes, direction) <- getMaxTrafficKRaw
    pure $ case direction of
      "down" -> hideNetDir $ MkNetBytesP @Down bytes
      "up" -> hideNetDir $ MkNetBytesP @Up bytes
      _ -> error "bad direction"
:}

We deliberately do not provide instances for SomeX classes that could be used to combine arbitrary SomeNetDirs (e.g. Applicative, AGroup), as that would defeat the purpose of enforcing the distinction between upload and downloaded bytes.

Equality is determined by the usual equivalence class -- that takes units into account -- and by considering the direction.

>>> hideNetDir (MkNetBytesP @Up @K 1000) == hideNetDir (MkNetBytesP @Up @K 1000)
True
>>> hideNetDir (MkNetBytesP @Up @K 1000) /= hideNetDir (MkNetBytesP @Down @K 1000)
True

Notice no Ord instance is provided, as we provide no ordering for Direction.

Since: 0.1

Constructors

MkSomeNetDir :: SDirection d -> NetBytes d s n -> SomeNetDir s n
Since: 0.1

hideNetDir :: forall d s n. SingDirection d => NetBytes d s n -> SomeNetDir s n

Wraps a NetBytes in an existentially quantified SomeNetDir.

Since: 0.1

someNetDirToSSize :: SingSize s => SomeNetDir s n -> SSize s

Retrieves the SingSize witness. Can be used to recover the Size.

Since: 0.1

Unknown Direction and Size

data SomeNet (n :: Type) where

Wrapper for NetBytes, existentially quantifying the size and direction. This is useful when a function does not know a priori what size or direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNet Double)
  getMaxTraffic = do
    -- getMaxTrafficRaw :: IO (Double, String, String)
    (bytes, direction, size) <- getMaxTrafficRaw
    pure $ case (direction, size) of
      ("Down", "K") -> hideNetSizeDir $ MkNetBytesP @Down @K bytes
      ("Up", "M") -> hideNetSizeDir $ MkNetBytesP @Up @M bytes
      _ -> error "todo"
:}

SomeNet carries along SDirection and SSize runtime witnesses for recovering the Direction and Size, respectively.

This instance uses the same equivalence relation from SomeNetSize w.r.t the size, and also includes an equality check on the direction. Thus we have, for instance,

>>> hideNetSizeDir (MkNetBytesP @Up @K 1_000) == hideNetSizeDir (MkNetBytesP @Up @M 1)
True

>>> hideNetSizeDir (MkNetBytesP @Up @K 1_000) /= hideNetSizeDir (MkNetBytesP @Down @M 1)
True

Since: 0.1

Constructors

MkSomeNet :: SDirection d -> SSize s -> NetBytes d s n -> SomeNet n
Since: 0.1

hideNetSizeDir :: forall d s n. (SingDirection d, SingSize s) => NetBytes d s n -> SomeNet n

Wraps a NetBytes in an existentially quantified SomeNet.

Since: 0.1