Data.Bytes.Size

Description

Provides the Size type and typeclasses for converting between units.

Since: 0.1

Size Tags

data Size

Byte units.

Since: 0.1

Constructors

B
Bytes

Since: 0.1

K
Kilobytes

Since: 0.1

M
Megabytes

Since: 0.1

G
Gigabytes

Since: 0.1

T
Terabytes

Since: 0.1

P
Petabytes

Since: 0.1

E
Exabytes

Since: 0.1

Z
Zettabytes

Since: 0.1

Y
Yottabytes

Since: 0.1

data SSize (s :: Size) where

Singleton for Size.

Since: 0.1

Constructors

SB :: SSize B
Since: 0.1

SK :: SSize K
Since: 0.1

SM :: SSize M
Since: 0.1

SG :: SSize G
Since: 0.1

ST :: SSize T
Since: 0.1

SP :: SSize P
Since: 0.1

SE :: SSize E
Since: 0.1

SZ :: SSize Z
Since: 0.1

SY :: SSize Y
Since: 0.1

class SingSize (s :: Size) where

Typeclass for recovering the Size at runtime.

Since: 0.1

Methods

singSize :: SSize s

Since: 0.1

withSingSize :: SSize s -> (SingSize s => r) -> r

Singleton "with"-style convenience function. Allows us to run a computation SingSize d => r without explicitly pattern-matching every time.

Since: 0.1

ssizeToSize :: SSize s -> Size

Since: 0.1

Sized Types

class Sized a where

Types that have a size.

Examples

Expand

>>> import Data.Bytes
>>> sizeOf (MkBytes @G 7)
G

>>> sizeOf (hideSize $ MkBytes @M 7)
M

Since: 0.1

Methods

sizeOf :: a -> Size

Retrieves the size.

Since: 0.1

Type Families for Relating Tags

type family NextSize (s :: Size) = (t :: Size) where ...

Type family that relates units to the next larger one.

Examples

Expand

>>> :kind! NextSize 'M
NextSize 'M :: Size
= 'G

>>> :kind! NextSize 'Y
NextSize 'Y :: Size
= (TypeError ...)

Since: 0.1

Equations

NextSize B = K

NextSize K = M

NextSize M = G

NextSize G = T

NextSize T = P

NextSize P = E

NextSize E = Z

NextSize Z = Y

NextSize Y = TypeError ('Text "The byte unit Y does not have a 'next size'.")

type family PrevSize (s :: Size) = (t :: Size) where ...

Type family that relates units to the previous smaller one.

Examples

Expand

>>> :kind! PrevSize 'M
PrevSize 'M :: Size
= 'K

>>> :kind! PrevSize 'B
PrevSize 'B :: Size
= (TypeError ...)

Since: 0.1

Equations

PrevSize B = TypeError ('Text "The byte unit B does not have a 'previous size'.")

PrevSize K = B

PrevSize M = K

PrevSize G = M

PrevSize T = G

PrevSize P = T

PrevSize E = P

PrevSize Z = E

PrevSize Y = Z